% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algr2_1.R
\docType{data}
\name{A11}
\alias{A11}
\title{general_CG}
\format{
An object of class \code{matrix} with 3 rows and 4 columns.
}
\usage{
A11
}
\arguments{
\item{A}{input matrixes}

\item{B}{input matrixes}
}
\value{
solve of matrix function
}
\description{
general_CG
}
\examples{
general_GC

}
\keyword{datasets}
