% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimCalculator.R
\name{dimCalculator}
\alias{dimCalculator}
\title{dimCalculator}
\usage{
dimCalculator(A, B, C)
}
\arguments{
\item{A}{input list consists of n^2 coupled matrixes A}

\item{B}{input list consists of n^2 coupled matrixes B}

\item{C}{input list consists of n coupled matrixes B}
}
\value{
return the dimension of X if the input is valid.
}
\description{
Give the list of input A,B and C check if the input is valid in dimension for martix group X, and return the
dim of X if it is valid
}
\examples{
dimCalculator(A,B,C)

}
